requirements.txt
flask
requests
python-whois
PyQt6